+++ src/components/ui/ContentTile/index.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 
 import { VideoTile } from '../../ui/VideoTile';
 
-export const ContentTile = styled(VideoTile)`
+export const ContentTile = styled<any>(VideoTile)`
   background-color: ${({ theme }) => theme.colors.greys.grey80};
 `;